EY-3488: Rydder litt opp i barnepensjon-fraser
