+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/etterlatte/maler/omstillingsstoenad/varsel/OmstillingsstoenadVarselRedigerbartUtfall.kt
@@ -10,7 +10,7 @@ import no.nav.pensjon.etterlatte.EtterlatteBrevKode
 import no.nav.pensjon.etterlatte.EtterlatteTemplate
 import no.nav.pensjon.etterlatte.maler.Delmal
 import no.nav.pensjon.etterlatte.maler.ManueltBrevDTO
-import no.nav.pensjon.etterlatte.maler.fraser.barnepensjon.RedigerbartUtfallFraser
+import no.nav.pensjon.etterlatte.maler.fraser.common.Felles
 
 @TemplateModelHelpers
 object OmstillingsstoenadVarselRedigerbartUtfall : EtterlatteTemplate<ManueltBrevDTO>, Delmal {
@@ -35,7 +35,7 @@ object OmstillingsstoenadVarselRedigerbartUtfall : EtterlatteTemplate<ManueltBre
             )
         }
         outline {
-            includePhrase(RedigerbartUtfallFraser.BlankTekst)
+            includePhrase(Felles.BlankTekst)
         }
     }
 }
